feat: update CanvasSwitcher.cs
CanvasSwitcher를 Singleton으로 바꾸었습니다.

환경설정을 수정하지 않더라도 기존 파일을 확인하도록 했습니다. 읽은 정보는 (settingsSave) 항상 들고 있습니다.

슬라이더와 관련된 변수들의 타입을 int로 바꾸었습니다. 이 부분을 따로 커밋하려 했는데, 충돌이 발생해서 줄별 커밋이 불가능했습니다 (얼마 전부터 자주 발생하는 문제인데, 원인은 잘 모르겠습니다).

+++ Gravity Controller/Assets/Scripts/Save/SettingsSave.cs
@@ -3,9 +3,9 @@ using UnityEngine;
 [SerializeField]
 public class SettingsSave
 {
-	public float backGroundVolume;
-	public float effectVolume;
-	public float sensitivity;
+	public int backGroundVolume;
+	public int effectVolume;
+	public int sensitivity;
 
 	public static SettingsSave Restore(string json)
 	{
